backend/home_backend/requirements.txt
asgiref==3.11.1
dj-database-url==3.1.0
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
drf-spectacular==0.29.0
djangorestframework-simplejwt==5.5.1
pillow==12.1.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
python-dateutil==2.9.0.post0
